Move add_ccpo_users functionality from database_setup.py into new script
